% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ms.R
\name{readMS}
\alias{readMS}
\title{Reading MS files

The \code{readMS} function imports the data from a file in *.ms format used by e.g. SIRIUS software reading
all specified fields and returning the data as a \code{\link[=DataFrame]{DataFrame()}}.}
\usage{
readMS(f, msLevel = 2L, mapping = spectraVariableMapping(MsBackendMS()), ...)
}
\arguments{
\item{f}{\code{character(1)} with the path to an MS file.}

\item{msLevel}{\code{numeric(1)} with the MS level. Default is 2. This value will
be reported as the spectra's MS level \strong{unless} the source MS file
defines the MS level.}

\item{mapping}{named \code{character} vector to rename MS fields to spectra
variables (see \code{\link[=spectraVariableMapping]{spectraVariableMapping()}}). This allows to correctly
import also custom fields or data from files with different MS
\emph{flavors}.}

\item{...}{Additional parameters, currently ignored.}
}
\value{
A \code{DataFrame} with each row containing the data from one spectrum
in the MS file. m/z and intensity values are available in columns \code{"mz"}
and \code{"intensity"} in a list representation.
}
\description{
Reading MS files

The \code{readMS} function imports the data from a file in *.ms format used by e.g. SIRIUS software reading
all specified fields and returning the data as a \code{\link[=DataFrame]{DataFrame()}}.
}
\examples{

fls <- dir(system.file("extdata", package = "MsBackendMS"),
    full.names = TRUE, pattern = "ms$")[1L]

readMsp(fls)
}
\author{
Laurent Gatto, Steffen Neumann, Johannes Rainer
}
